feat: modularizaçao da funcao carregar alunos
